Removed duplicate PullRequestListFragment
