+++ app/src/main/java/io/jeffchang/splitdiff/di/module/FragmentBuilderModule.kt
@@ -2,7 +2,7 @@ package io.jeffchang.splitdiff.di.module
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import io.jeffchang.splitdiff.ui.gitdiff.PullRequestListFragment
+import io.jeffchang.splitdiff.ui.pullrequests.PullRequestListFragment
 
 @Suppress("unused")
 @Module
