+++ app/src/main/java/io/jeffchang/splitdiff/ui/pullrequests/PullRequestListFragment.kt
@@ -4,10 +4,17 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import dagger.android.support.DaggerFragment
 import io.jeffchang.splitdiff.R
+import io.jeffchang.splitdiff.ui.gitdiff.viewmodel.PullRequestViewModel
+import kotlinx.android.synthetic.main.fragment_pull_request_list.*
+import javax.inject.Inject
 
-class PullRequestListFragment: Fragment() {
+class PullRequestListFragment: DaggerFragment() {
+
+    @Inject
+    lateinit var pullRequestViewModel: PullRequestViewModel
 
     override fun onCreateView(
             inflater: LayoutInflater,
@@ -17,6 +24,18 @@ class PullRequestListFragment: Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        subscribeUi()
+        pullRequestViewModel.getPullRequests()
+    }
+
+    private fun subscribeUi() {
+        pullRequestViewModel.textDataLiveData.observe(this, Observer {
+            fragment_pull_request_loading_view.textData = it
+        })
+
+        pullRequestViewModel.pullRequestLiveData.observe(this, Observer {
+
+        })
     }
 
 }
