+++ app/src/main/res/navigation/nav_pull_request.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/pullRequestFragment"
-        android:name="io.jeffchang.splitdiff.ui.gitdiff.PullRequestListFragment"
+        android:name="io.jeffchang.splitdiff.ui.pullrequests.PullRequestListFragment"
         android:label="WordSearchFragment"
         tools:layout="@layout/fragment_pull_request_list" />
